[FIX] 'custom_account': Mejorada implementacion y jerarquia del proceso de envio de correos automatico
